Issue #81: Move the post-login event listener in the new module.
