+++ src/Event/EuLoginEventSubscriber.php
@@ -4,7 +4,6 @@ declare(strict_types = 1);
 
 namespace Drupal\oe_authentication\Event;
 
-use Drupal\cas\Event\CasPostLoginEvent;
 use Drupal\cas\Event\CasPostValidateEvent;
 use Drupal\cas\Event\CasPreRegisterEvent;
 use Drupal\cas\Event\CasPreValidateEvent;
@@ -55,7 +54,8 @@ class EuLoginEventSubscriber implements EventSubscriberInterface {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('config.factory')
+      $container->get('config.factory'),
+      $container->get('messenger')
     );
   }
 
@@ -67,29 +67,10 @@ class EuLoginEventSubscriber implements EventSubscriberInterface {
    */
   public static function getSubscribedEvents(): array {
     $events = [];
-    $events[CasHelper::EVENT_POST_LOGIN] = 'updateUserData';
     $events[CasHelper::EVENT_PRE_REGISTER] = 'processUserProperties';
     $events[CasHelper::EVENT_POST_VALIDATE] = 'processCasAttributes';
     $events[CasHelper::EVENT_PRE_VALIDATE] = 'alterValidationPath';
     return $events;
-  }
-
-  /**
-   * Updates the user data based on the information taken from EU Login.
-   *
-   * @param \Drupal\cas\Event\CasPostLoginEvent $event
-   *   The triggered event.
-   *
-   * @throws \Drupal\Core\Entity\EntityStorageException
-   *   In case of failures an exception is thrown.
-   */
-  public function updateUserData(CasPostLoginEvent $event): void {
-    $properties = CasProcessor::convertCasAttributesToFieldValues($event->getCasPropertyBag()->getAttributes());
-    $account = $event->getAccount();
-    foreach ($properties as $name => $value) {
-      $account->set($name, $value);
-    }
-    $account->save();
   }
 
   /**
@@ -99,10 +80,6 @@ class EuLoginEventSubscriber implements EventSubscriberInterface {
    *   The triggered event.
    */
   public function processUserProperties(CasPreRegisterEvent $event): void {
-
-    $attributes = $event->getCasPropertyBag()->getAttributes();
-    $event->setPropertyValues(CasProcessor::convertCasAttributesToFieldValues($attributes));
-
     // If the site is configured to need administrator approval,
     // change the status of the account to blocked.
     $user_settings = $this->configFactory->get('user.settings');
